Refactor cluster election and transport layers
- Split shared node logic into cluster, election, protocol, transport, and testing modules
- Update KV example and action sender to handle no-leader errors
- Bump crate version to 1.0.0
