+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sharedstate"
-version = "0.5.1"
+version = "1.0.0"
 edition = "2021"
 authors = ["Patrick Lorio <patrick@methods.dev>"]
 description = "Sync heavily read state across many servers"
