+++ src/cluster/mod.rs
@@ -1,0 +1,11 @@
+//! Cluster coordination and node facade.
+
+pub mod action_pump;
+pub mod control;
+pub mod election;
+pub mod leader;
+pub mod node;
+pub mod observer;
+pub mod peer_handler;
+
+pub use node::*;
