+++ src/protocol/mod.rs
@@ -1,0 +1,4 @@
+//! Wire protocol messages and binary framing.
+
+pub mod framing;
+pub mod messages;
