+++ src/transport/mod.rs
@@ -1,0 +1,5 @@
+//! Network transport traits, channel workers, and simulated transport.
+
+pub mod channels;
+pub mod simulated;
+pub mod traits;
