+++ src/utils.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use std::{
     fmt::Debug,
     future::Future,
